+++ docs/reference/bigframes.ml/impute.rst
@@ -1,0 +1,7 @@
+bigframes.ml.impute
+==========================
+
+.. automodule:: bigframes.ml.impute
+    :members:
+    :inherited-members:
+    :undoc-members:
