+++ tests/system/small/ml/test_preprocessing.py
@@ -14,13 +14,11 @@
 
 import math
 
-import numpy as np
 import pandas as pd
 import pyarrow as pa
 
 import bigframes.features
 from bigframes.ml import preprocessing
-import bigframes.pandas as bpd
 
 ONE_HOT_ENCODED_DTYPE = (
     pd.ArrowDtype(pa.list_(pa.struct([("index", pa.int64()), ("value", pa.float64())])))
@@ -544,99 +542,6 @@ def test_k_bins_discretizer_save_load_quantile(new_penguins_df, dataset_id):
     )
     assert isinstance(reloaded_transformer, preprocessing.KBinsDiscretizer)
     assert reloaded_transformer.n_bins == transformer.n_bins
-    assert reloaded_transformer.strategy == transformer.strategy
-    assert reloaded_transformer._bqml_model is not None
-
-
-def test_simple_imputer_normalized_fit_transform_default_params():
-    missing_df = bpd.DataFrame(
-        {
-            "culmen_length_mm": [39.5, 38.5, 37.9],
-            "culmen_depth_mm": [np.nan, 17.2, 18.1],
-            "flipper_length_mm": [np.nan, 181.0, 188.0],
-        }
-    )
-    imputer = preprocessing.SimpleImputer(strategy="mean")
-    result = imputer.fit_transform(
-        missing_df[["culmen_length_mm", "culmen_depth_mm", "flipper_length_mm"]]
-    ).to_pandas()
-
-    expected = pd.DataFrame(
-        {
-            "imputer_culmen_length_mm": [39.5, 38.5, 37.9],
-            "imputer_culmen_depth_mm": [17.65, 17.2, 18.1],
-            "imputer_flipper_length_mm": [184.5, 181.0, 188.0],
-        },
-        dtype="Float64",
-        index=pd.Index([0, 1, 2], dtype="Int64"),
-    )
-
-    pd.testing.assert_frame_equal(result, expected)
-
-
-def test_simple_imputer_series_normalizes(new_penguins_df):
-    missing_df = bpd.DataFrame(
-        {
-            "culmen_length_mm": [39.5, 38.5, 37.9],
-            "culmen_depth_mm": [np.nan, 17.2, 18.1],
-            "flipper_length_mm": [np.nan, 181.0, 188.0],
-        }
-    )
-    imputer = preprocessing.SimpleImputer()
-    imputer.fit(missing_df["culmen_depth_mm"])
-
-    result = imputer.transform(missing_df["culmen_depth_mm"]).to_pandas()
-    result = imputer.transform(new_penguins_df).to_pandas()
-
-    expected = pd.DataFrame(
-        {
-            "imputer_culmen_depth_mm": [18.8, 17.2, 18.1],
-        },
-        dtype="Float64",
-        index=pd.Index([1633, 1672, 1690], name="tag_number", dtype="Int64"),
-    )
-
-    pd.testing.assert_frame_equal(result, expected, rtol=0.1)
-
-
-def test_simple_imputer_save_load_mean(dataset_id):
-    missing_df = bpd.DataFrame(
-        {
-            "culmen_length_mm": [39.5, 38.5, 37.9],
-            "culmen_depth_mm": [np.nan, 17.2, 18.1],
-            "flipper_length_mm": [np.nan, 181.0, 188.0],
-        }
-    )
-    transformer = preprocessing.SimpleImputer(strategy="mean")
-    transformer.fit(
-        missing_df[["culmen_length_mm", "culmen_depth_mm", "flipper_length_mm"]]
-    )
-
-    reloaded_transformer = transformer.to_gbq(
-        f"{dataset_id}.temp_configured_model", replace=True
-    )
-    assert isinstance(reloaded_transformer, preprocessing.SimpleImputer)
-    assert reloaded_transformer.strategy == transformer.strategy
-    assert reloaded_transformer._bqml_model is not None
-
-
-def test_simple_imputer_save_load_most_frequent(dataset_id):
-    missing_df = bpd.DataFrame(
-        {
-            "culmen_length_mm": [39.5, 38.5, 37.9],
-            "culmen_depth_mm": [np.nan, 17.2, 18.1],
-            "flipper_length_mm": [np.nan, 181.0, 188.0],
-        }
-    )
-    transformer = preprocessing.SimpleImputer(strategy="most_frequent")
-    transformer.fit(
-        missing_df[["culmen_length_mm", "culmen_depth_mm", "flipper_length_mm"]]
-    )
-
-    reloaded_transformer = transformer.to_gbq(
-        f"{dataset_id}.temp_configured_model", replace=True
-    )
-    assert isinstance(reloaded_transformer, preprocessing.SimpleImputer)
     assert reloaded_transformer.strategy == transformer.strategy
     assert reloaded_transformer._bqml_model is not None
